+++ tests/builders/primary-term-builder-test.php
@@ -8,7 +8,7 @@ use Yoast\WP\SEO\Helpers\Meta_Helper;
 use Yoast\WP\SEO\Helpers\Primary_Term_Helper;
 use Yoast\WP\SEO\Repositories\Primary_Term_Repository;
 use Yoast\WP\SEO\Tests\Doubles\Builders\Primary_Term_Builder_Double;
-use Yoast\WP\SEO\Tests\Mocks\Primary_Term_Mock;
+use Yoast\WP\SEO\Tests\Doubles\Models\Primary_Term_Mock;
 use Yoast\WP\SEO\Tests\TestCase;
 
 /**
